fix: make swaggerui document available through API and through UI interface
